NullShellcode/requirements.txt
click>=8.0.0
keystone-engine>=0.9.2
capstone>=4.0.2
pytest>=7.0.0  